requirements.txt
Django==1.6
MySQL-python==1.2.5
PIL==1.1.7
South==0.8.4
argparse==1.2.1
django-allauth==0.15.0
django-haystack==2.1.0
oauthlib==0.6.3
psutil==1.2.1
pysolr==3.2.0
python-openid==2.2.5
python-ptrace==0.7
requests==2.3.0
requests-oauthlib==0.4.1
simplejson==3.5.3
six==1.7.3
urllib3==1.9
wsgiref==0.1.2
